MVP: blue print reformat
