+++ .pylintrc
@@ -44,7 +44,11 @@ fail-under=10
 # Files or directories to be skipped. They should be base names, not paths.
 ignore=CVS,
       venv,
-      instance
+      instance,
+      config,
+      tests,
+      htmlcov,
+      static
 
 # Add files or directories matching the regex patterns to the ignore-list. The
 # regex matches against paths and can be in Posix or Windows format.
