+++ old/clearFormsButton.html
@@ -1,0 +1,2 @@
+<button class="btn btn-outline-info" type="submit"
+formaction="{{ url_for('clear_form_route', form_name='RegistrationForm') }}">Clear</button>
