+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.black]
-line-length = 80
+line-length = 88
 target-version = ['py38']
 [tool.black.docstring-style]
-wrap = 80
+wrap = 88
