--- weight_tracker/constants/choices.py
@@ -1,11 +1,0 @@
-# choices.py
-TIME_OF_DAY_CHOICES = [("day", "Day"), ("night", "Night")]
-
-MOOD_CHOICES = [
-    ("good", "Good"),
-    ("average", "Average"),
-    ("below_average", "Below Average"),
-    ("terrible", "Terrible"),
-]
-
-STATUS_CHOICES = [("normal", "Normal"), ("sick", "Sick"), ("injured", "Injured")]
