+++ weight_tracker/src/constants/choices.py
@@ -1,0 +1,11 @@
+"""Contants module containing choices for drop down selectors"""
+TIME_OF_DAY_CHOICES = [("day", "Morning"), ("night", "Evening"), ("afternoon", "Afternoon")]
+
+MOOD_CHOICES = [
+    ("happy", "Happy"),
+    ("average", "Average"),
+    ("below_average", "Depressed"),
+    ("terrible", "Tired"),
+]
+
+STATUS_CHOICES = [("normal", "Normal"), ("sick", "Sick"), ("injured", "Injured")]
