+++ weight_tracker/src/static/default.jpg
