+++ weight_tracker/src/templates/entry.html
@@ -46,8 +46,8 @@
                         </div>
                         <div class="d-flex">
                             <a class="btn btn-primary ml-auto mr-2"
-                                href="{{ url_for('update_entry', entry_id=entry.id, entry=entry) }}">Edit</a>
-                            <form method="post" action="{{ url_for('delete_entry', entry_id=entry.id) }}">
+                                href="{{ url_for('entries.update_entry', entry_id=entry.id, entry=entry) }}">Edit</a>
+                            <form method="post" action="{{ url_for('entries.delete_entry', entry_id=entry.id) }}">
                                 <button type="submit" class="btn btn-danger">Delete</button>
                             </form>
                         </div>
