+++ weight_tracker/src/templates/entry_edit.html
@@ -32,7 +32,7 @@
             <div class="d-flex flex-column">
                 <div class="card mb-12">
                     <div class="card-body">
-                        <form method="POST" action="{{ url_for('update_entry', entry_id=entry.id) }}">
+                        <form method="POST" action="{{ url_for('entries.update_entry', entry_id=entry.id) }}">
                             <input type="hidden" name="_method" value="PUT">
                             {{ form.csrf_token }}
                             <div class="form-group">
@@ -100,7 +100,7 @@
                                 {% endfor %}
                             </div>
                             {{ form.submit(class="btn btn-outline-info") }}
-                            <a href="{{ url_for('get_entry', entry_id=entry.id) }}" class="btn btn-secondary">Cancel</a>
+                            <a href="{{ url_for('entries.get_entry', entry_id=entry.id) }}" class="btn btn-secondary">Cancel</a>
                         </form>
                     </div>
                 </div>
