+++ weight_tracker/src/templates/login.html
@@ -60,12 +60,12 @@
   <div class="border-top pt-3">
     <p>
       <small class="text-muted">
-      Need An Account? <a href="{{ url_for('register') }}">Sign Up Now</a>
+      Need An Account? <a href="{{ url_for('users.register') }}">Sign Up Now</a>
       </small>
     </p>
     <p>
       <small class="text-muted">
-      Click <a href="{{ url_for('logout') }}">here</a> to test the logout route.
+      Click <a href="{{ url_for('users.logout') }}">here</a> to test the logout route.
       </small>
     </p>
   </div>
