+++ weight_tracker/src/users/routes.py
@@ -1,0 +1,98 @@
+"""
+This is a Flask blueprint module that contains routes related to user
+functionality.
+
+Routes:
+/login: renders a login form and handles user authentication.
+/register: renders a registration form and handles adding users.
+/logout: logs current user out of current session.
+/account: renders user information.
+"""
+from flask import render_template, redirect, url_for, flash, Blueprint
+from flask_login import login_required, login_user, current_user, logout_user
+from weight_tracker.src import db, bcrypt
+from weight_tracker.src.users.forms import LoginForm, RegistrationForm
+from weight_tracker.src.models import User
+
+users = Blueprint("users", __name__)
+
+
+@users.route("/login", methods=["POST", "GET"])
+def login():
+    """
+    If the form is submitted and the email and password are correct, then the
+    user is logged in and redirected to the tracker page
+    :return: The login.html template is being returned.
+    """
+    if current_user.is_authenticated:
+        print(vars(current_user))
+        return redirect(url_for("main.home"))
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data, active_record=True).first()
+        if user and bcrypt.check_password_hash(user.password, form.password.data):
+            login_user(user, remember=form.remember.data)
+            flash("Login Successful", "success")
+            return redirect(url_for("main.home"))
+        flash("Login Failed. Please check username/password", "danger")
+    return render_template("login.html", active_page="login", form=form)
+
+
+@users.route("/logout")
+def logout():
+    """
+    Route: logs out current user.
+    :return: the redirect function, which is redirecting the user to the home.
+    """
+    logout_user()
+    print("log this: user logged out")
+    return redirect(url_for("main.home"))
+
+
+@users.route("/register", methods=["GET", "POST"])
+def register():
+    """
+    The function takes in a form object, validates the form, hashes the
+    password, creates a new user object, adds the new user to the database,
+    commits the changes to the database, flashes a message to the user, and
+    redirects the user to the login page.
+
+    Bugs: Experienced a bug where after registering, user was automatically
+    logged in, instead of being redircted to the login in page. I think what
+    might of happened has to do with cache. I was logged in with a user, deleted
+    the database file. used a create_all(), then opened an icognito window, then
+    registered with the same username/credentials as the user I was previously
+    logged in with. Check on this later. For now will add logout() before
+    redirect for safety.
+    :return: the rendered template for the register page.
+    """
+    if current_user.is_authenticated:
+        return redirect(url_for("main.home"))
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        hashed_pw = bcrypt.generate_password_hash(form.password.data).decode("utf8")
+        new_user = User(
+            username=form.username.data,
+            email=form.email.data,
+            password=hashed_pw,
+        )
+        db.session.add(new_user)
+        db.session.commit()
+        logout_user()
+        flash("Your account has been created. You may now login.", "success")
+        return redirect(url_for("users.login"))
+    return render_template("register.html", active_page="register", form=form)
+
+
+@users.route("/account", methods=["GET"])
+@login_required
+def account():
+    """
+    Renders the account page template and passes the current user's data to the template.
+    If the user is not authenticated, redirects to the login page.
+
+    :return: The account.html template with the current user's data.
+    """
+    if current_user.is_authenticated:
+        return render_template("account.html", active_page="account", user=current_user)
+    return redirect(url_for("users.login"))
