+++ weight_tracker/src/utils/helpers.py
@@ -1,0 +1,7 @@
+"""Helpers module for flask application, weight-tracker"""
+
+
+def clear_form(form):
+    """NOT IMPLEMENTED. Clears all fields in the given form."""
+    # form.process(data=None)
+    print(f"clearing fields from {form}")
